Breath comma and redundant comma are removed
The phrase "from a monolith, to separate services, and then back" has two
unnecessary commas. From a monolith to separate services is fine without it
and the comma seems to be added as a breath or pause marker. The second comma
is made redundant by the "and", since it's not in a list. It would be necessary
if the phrase read "from a monolith to separate services, then back," though.